+++ lib/ruql.rb
@@ -22,6 +22,6 @@ require 'ruql/multiple_choice'
 require 'ruql/select_multiple'
 require 'ruql/true_false'
 require 'ruql/fill_in'
-require 'ruql/peer_review/peer_review'
-require 'ruql/peer_review/criterion'
-require 'ruql/peer_review/option'
+require 'ruql/open_assessment/open_assessment'
+require 'ruql/open_assessment/criterion'
+require 'ruql/open_assessment/option'
