+++ lib/ruql/open_assessment/criterion.rb
